requirements.txt
Flask==2.3.3
tensorflow==2.19.0
Pillow==10.1.0
flask-cors==4.0.0
gunicorn==21.2.0
numpy==1.26.0
